+++ lib/commanded/commands/router.ex
@@ -8,7 +8,18 @@ defmodule Commanded.Commands.Router do
     end
   end
 
-  # dispatch the given command to the corresponding handler for a given aggregate root uniquely identified
+  @doc """
+  Dispatch the given command to the corresponding handler for a given aggregate root uniquely identified
+  """
+  defmacro dispatch(command_module, opts)
+
+  defmacro dispatch(command_modules, opts) when is_list(command_modules) do
+    quote do
+      unquote(command_modules)
+      |> Enum.map(fn command_module -> dispatch(command_module, unquote(opts)) end)
+    end
+  end
+
   defmacro dispatch(command_module, to: handler, aggregate: aggregate, identity: identity) do
     quote do
       if Enum.member?(@registered_commands, unquote(command_module)) do
@@ -22,6 +33,7 @@ defmodule Commanded.Commands.Router do
 
       Returns `:ok` on success.
       """
+      def dispatch(command)
       def dispatch(%unquote(command_module){} = command) do
         Commanded.Commands.Dispatcher.dispatch(command, unquote(handler), unquote(aggregate), unquote(identity))
       end